Rename MainMenu to MainPagee
