+++ client/src/components/Main/LogoutButton.js
@@ -1,4 +1,4 @@
-import styles from './MainMenu.module.css';
+import styles from './MainPage.module.css';
 const LogoutButton = ({ onLogout }) => {
     return (
         <button className={styles.logoutButton} onClick={onLogout}>
